Basic CalendarCtrl Implementation
The baseline implementation of the controller gets the list of events
via EventService, and generates the dateRange var for each event to
simplify template implementation.

The dateRage == start day + ‘-‘ + end day, where ‘day’ is the day of
the month 1-31. If the end months are different, then the dateRange is
just the start day (no end day included).